[Model] Nemotron ASR: PrecisionPolicy delegation in compute (closes the beta2 gap)
The beta2 pre-pod audit found PORT-PREC-001 only half-realized: the
policy existed but no compute module consulted it — load_core built
fp32 unconditionally and a bf16 policy value changed nothing. This
lands the delegation so a precision change is again a policy value:

- apply_policy_dtypes(core): encoder/LID/predictor/joint move to
  dtype_for("weights"); the mel front-end stays fp32 (upstream of the
  activations seam, the golden input boundary); load_core_from_dump
  takes the policy and applies it after weight load.
- Entry/seam casts, each following the policy-cast weights rather than
  naming a dtype: subsampling entry (fp32 mel -> weight dtype),
  pos-emb refs take the activation dtype, Joint.logits bridges the
  fp32 recurrent-dtype pred_out to the weight dtype (PORT-PREC-005:
  the manual cell already up-casts weights to the state dtype, so
  (h,c) accumulate fp32 under bf16 weights).
- State classes are independent axes at every cache: StreamingCaches
  channel/time and the alpha pages read-cast to compute dtype and
  write-cast back, so pages may store wider than compute.
- BF16_COMPUTE: the first reviewed sub-fp32 policy value
  (PORT-PREC-003) — weights+activations bf16, every state class fp32,
  the vLLM mamba mixed-precision idiom. Transcript parity stays
  blocking; tensor deltas advisory under its identifier.

test_precision_lane (7/7): delegation mechanics, all mixed-dtype
seams, fp32-state-under-bf16-weights, activations!=weights refusal,
featurizer exclusion. Alpha suites re-run green (16/16); fp32 path
untouched by construction (casts are no-ops at fp32).

AI-assisted (Claude).

+++ vllm_omni/model_executor/models/nemotron_asr/attention_pages.py
@@ -122,8 +122,11 @@ def paged_stream_attention(
         new_frames=new_frames,
         window=window,
     )
-    k = k.transpose(1, 2)
-    v = v.transpose(1, 2)
+    # attention_cache is its own policy axis: pages may store a wider
+    # dtype than compute; read-cast to the activations dtype (writes
+    # cast implicitly on scatter).
+    k = k.to(x.dtype).transpose(1, 2)
+    v = v.to(x.dtype).transpose(1, 2)
     t2 = window + new_frames
 
     q = attn.linear_q(x).view(batch, new_frames, attn.h, attn.d_k)
